Refactor loader dependencies
Isolates the file loader so it does no longer depend on the parser.
This frees the loader to only care aout loading files, which makes it
easier to inject custom loaders via the provided interface.

Since the loader is now only a single function that is extracted
immediately, it's been reduced to a function instead of a class.